pyproject.toml
[project]
name = "pasteprompt"
version = "0.1.0"
description = "Quick prompt snippets for macOS Services menu"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Matt Jenior" }
]
keywords = ["macos", "automation", "prompts", "clipboard", "services"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]
menubar = [
    "rumps>=0.4.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-ApplicationServices>=10.0",
    "watchdog>=3.0.0",
]
all = [
    "pasteprompt[dev,menubar]",
]

[project.scripts]
pasteprompt = "pasteprompt.cli:cli"

[project.urls]
Homepage = "https://github.com/mjenior/pasteprompt"
Documentation = "https://github.com/mjenior/pasteprompt#readme"
Repository = "https://github.com/mjenior/pasteprompt.git"
Issues = "https://github.com/mjenior/pasteprompt/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pasteprompt"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]

